refactor: Renamed DataTab to PlotsTab, added hierarchy for toolbar
