+++ src/components/Tabs/Plot3d/Plot3dTab.tsx
@@ -5,20 +5,20 @@ import { useDebounce } from "usehooks-ts";
 import { DataTabType, SelectionOutlineColorMode, TabType, type VectorFieldData } from "src/colorizer";
 import { getSharedWorkerPool } from "src/colorizer/workers/SharedWorkerPool";
 import LoadingSpinner from "src/components/LoadingSpinner";
-import DataTabToolbar from "src/components/Tabs/Data/DataTabToolbar";
-import type { SharedDataTabProps } from "src/components/Tabs/Data/types";
 import Plot3dToolbar from "src/components/Tabs/Plot3d/Plot3dToolbar";
+import DataTabToolbar from "src/components/Tabs/Plots/PlotsTabToolbar";
+import type { SharedPlotTabProps } from "src/components/Tabs/Plots/types";
 import { useInteractionListener } from "src/hooks";
 import { useViewerStateStore } from "src/state";
 import { FlexColumn } from "src/styles/utils";
 
-import { make3dConeTrace } from "./plot_3d_utils";
 import Plot3d from "./Plot3d";
+import { make3dConeTrace } from "./plot_3d_utils";
 
 const MINIMUM_BIN_COUNT = 10;
 const RESUME_PLAYBACK_TIMEOUT_MS = 500;
 
-type Plot3dTabProps = SharedDataTabProps;
+type Plot3dTabProps = SharedPlotTabProps;
 
 export default function Plot3dTab(props: Plot3dTabProps): ReactElement {
   const plotContainerRef = useRef<HTMLDivElement>(null);
@@ -217,8 +217,8 @@ export default function Plot3dTab(props: Plot3dTabProps): ReactElement {
   return (
     <FlexColumn style={{ height: "100%", marginBottom: 10 }} $gap={8}>
       <DataTabToolbar>
-        <Plot3dToolbar />
         {props.toolbar}
+        <Plot3dToolbar />
       </DataTabToolbar>
 
       {/* Plot Container */}
