+++ src/components/Tabs/index.ts
@@ -1,6 +1,6 @@
 import AnnotationTab from "./Annotation/AnnotationTab";
-import DataTab from "./Data/DataTab";
 import FeatureThresholdsTab from "./Filters/FeatureThresholdsTab";
+import PlotsTab from "./Plots/PlotsTab";
 import SettingsTab from "./Settings/SettingsTab";
 
-export { AnnotationTab, DataTab, FeatureThresholdsTab, SettingsTab };
+export { AnnotationTab, PlotsTab, FeatureThresholdsTab, SettingsTab };
